Refactor API files: update import paths for models and dependencies, remove unused database dependencies, and delete obsolete model definitions for improved organization and maintainability.
