+++ app/api/metadata.py
@@ -13,5 +13,3 @@ async def get_metadata(url: AnyUrl = Query(..., description="Target URL to fetch
         return fetch_metadata_logic(str(url))
     except Exception as e:
         raise HTTPException(status_code=500, detail=f"Failed to fetch metadata: {str(e)}")
-# Use shared dependencies (if needed for future DB access)
-from app.database.dependencies import db_dependency, user_dependency
