--- app/db/database.py
@@ -1,23 +1,0 @@
-import os
-from dotenv import load_dotenv
-
-load_dotenv()
-
-from sqlalchemy import create_engine
-from sqlalchemy.orm import sessionmaker
-from sqlalchemy.ext.declarative import declarative_base
-from psycopg2 import sql
-
-
-DATABASE_URL = os.environ.get("DATABASE_URL")
-
-if not DATABASE_URL:
-    DATABASE_URL = 'postgresql://postgres:060204@localhost/Linkify'
-
-
-engine = create_engine(DATABASE_URL)
-
-SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
-
-Base = declarative_base()
-
